requirements.txt
Flask==3.0.3 
python-dotenv==1.0.1 
Flask-RESTful==0.3.10 
Flask-SQLAlchemy==3.1.1 
Flask-JWT-Extended==4.6.0 
Flask-Mail==0.10.0
Flask-Testing==0.8.1 
pytest==8.2.2
requests==2.32.3